created separate Logo component
